# docs/openapi.yaml
openapi: 3.1.0
info:
  title: AI Team Hub API
  description: |
    AI Team Hub Public API — Multi-Agent AI Runtime.

    Three modes available:
    - **auto** (default): Full FSM + agents + cache pipeline
    - **control**: User overrides agent behavior
    - **debug**: Full trace, FSM states, internal logs

    All endpoints return a unified response schema.
  version: "2.1.0"
  contact:
    name: AI Team Hub
  license:
    name: MIT

servers:
  - url: https://ai-team-hub.wt5371.workers.dev
    description: Production (Cloudflare Workers)
  - url: http://localhost:8910
    description: Local / Docker

tags:
  - name: v1-public
    description: Public API (requires X-API-Key)
  - name: v1-observability
    description: Timeline, graphs, cache visualization (requires X-API-Key)

components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key

  schemas:
    TaskResponse:
      type: object
      properties:
        task_id: { type: string }
        status: { type: string }
        result: { type: string }
        trace_id: { type: string }
        cost: { type: string, example: "$0" }
        latency: { type: string, example: "1234ms" }
        message: { type: string }

    WorkspaceResponse:
      type: object
      properties:
        workspace_id: { type: string }
        status: { type: string }
        title: { type: string }
        created_at: { type: string }
        latency: { type: string }

    TraceResponse:
      type: object
      properties:
        trace_id: { type: string }
        task_id: { type: string }
        status: { type: string }
        steps: { type: array, items: { type: object } }
        fsm_transitions:
          type: array
          items:
            type: object
            properties:
              from: { type: string }
              to: { type: string }
        agent_calls: { type: array, items: { type: object } }
        cache_hits: { type: integer }
        total_cost: { type: string }
        total_latency: { type: string }
        message: { type: string }

    ChatResponse:
      type: object
      properties:
        session_id: { type: string }
        status: { type: string }
        response: { type: string }
        agent_used: { type: string }
        latency: { type: string }
        message: { type: string }

security:
  - ApiKeyAuth: []

paths:
  /v1/task/run:
    post:
      tags: [v1-public]
      summary: Execute a task
      description: Execute a task through the AI Runtime with chosen mode (auto/control/debug).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [task]
              properties:
                task: { type: string, description: "Task description" }
                mode:
                  type: string
                  enum: [auto, control, debug]
                  default: auto
                provider: { type: string, default: openrouter }
                model: { type: string, default: "openrouter/owl-alpha" }
                budget: { type: number, default: 0.5 }
                timeout: { type: integer, default: 120 }
                agent_config:
                  type: object
                  description: "Control mode: override agent behavior"
                workspace_id: { type: string }
      responses:
        "200":
          description: Task response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskResponse"
        "401":
          description: Missing API key
        "403":
          description: Invalid API key

  /v1/workspace/create:
    post:
      tags: [v1-public]
      summary: Create workspace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title: { type: string }
                description: { type: string }
      responses:
        "200":
          description: Workspace created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceResponse"

  /v1/task/{task_id}/status:
    get:
      tags: [v1-public]
      summary: Get task status
      parameters:
        - name: task_id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Task status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskResponse"

  /v1/task/{task_id}/trace:
    get:
      tags: [v1-public]
      summary: Get execution trace
      parameters:
        - name: task_id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Full trace data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TraceResponse"

  /v1/agent/chat:
    post:
      tags: [v1-public]
      summary: Simple agent chat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message]
              properties:
                message: { type: string }
                session_id: { type: string }
                mode: { type: string, default: auto }
                context: { type: object }
      responses:
        "200":
          description: Chat response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"

  /v1/health:
    get:
      tags: [v1-public]
      summary: Health check
      security: []
      responses:
        "200":
          description: Service healthy

  /v1/system/modes:
    get:
      tags: [v1-public]
      summary: Available modes
      security: []
      responses:
        "200":
          description: List of modes

  /v1/timeline/{task_id}:
    get:
      tags: [v1-observability]
      summary: Task timeline for UI
      parameters:
        - name: task_id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Timeline data

  /v1/agent-graph/{task_id}:
    get:
      tags: [v1-observability]
      summary: Agent execution graph
      parameters:
        - name: task_id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Graph nodes and edges

  /v1/cost/{task_id}:
    get:
      tags: [v1-observability]
      summary: Cost breakdown
      parameters:
        - name: task_id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Cost breakdown

  /v1/cache/vis:
    get:
      tags: [v1-observability]
      summary: Cache visualization
      responses:
        "200":
          description: Cache hit/miss data

  /v1/fsm-transitions/{task_id}:
    get:
      tags: [v1-observability]
      summary: FSM state transitions
      parameters:
        - name: task_id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: FSM transition log